rtfd_requirements.txt
petl>=1.0
numpydoc==0.5
